requirements.txt
asgiref==3.5.0
cx-Oracle==8.3.0
Django==4.0.3
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.1
